+++ config/config.json
@@ -8,10 +8,10 @@
     "connect_selenium_grid": "false",
     "supervision_password": "patladjan1",
     "wda_bundle_id": "com.shamanec.WebDriverAgentRunner.xctrunner",
-    "host_address": "192.168.1.13",
-    "mongo_db": "192.168.1.13:27017",
+    "host_address": "192.168.1.6",
+    "mongo_db": "192.168.1.28:27017",
     "wda_repo_path": "/Users/shamanec/Downloads/WebDriverAgent-5.8.3/",
-    "provider_nickname": "Provider1"
+    "provider_nickname": "Provider2"
   },
   "devices-config": [
     {
