+++ config/device-sync.sh
